Change database key types from int to bigint
Manually dropped and re-added constraints in migration.

+++ SteamMates/Models/Persistence/Rating.cs
@@ -4,17 +4,17 @@ namespace SteamMates.Models.Persistence
 {
     public class Rating
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public UserIdentifier User { get; set; }
 
         [ForeignKey(nameof(User))]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         public GameIdentifier Game { get; set; }
 
         [ForeignKey(nameof(Game))]
-        public int GameId { get; set; }
+        public long GameId { get; set; }
 
         public int Value { get; set; }
     }
